Adds org-pomodoro layer
@swaroopch already made the layer and it has been added to the next
release of syl20bnr/spacemacs it seems. (It is on the develop branch at
the time of writing this)

The pull request is at https://github.com/syl20bnr/spacemacs/pull/934

I have just shamelesly copied the code. I will remove this layer from my
private layers and use the official one, as soon as that release is out.